# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Barbossa API
  description: |
    Family music library management API.

    ## Authentication
    All endpoints except `/api/auth/login` require a valid JWT token.
    Pass token in Authorization header: `Bearer <token>`

    ## User Roles
    - **Regular**: Can browse, heart, download, export own library
    - **Admin**: Can manage users, delete from disk, upload to TorrentLeech
  version: 0.1.9
  contact:
    name: Barbossa
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://barbossa:8080
    description: Docker internal
...
paths:
...
  /api/me/export:
    post:
      tags: [library]
      summary: Export user library
      operationId: exportLibrary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '202':
          description: Export started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportStatus'
  /api/me/export/{id}:
    get:
      tags: [library]
      summary: Get export status
      operationId: getExportStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Export status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportStatus'
...
  /api/downloads/search/qobuz:
    get:
      tags: [downloads]
      summary: Search Qobuz
      operationId: searchQobuz
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [artist, album, track, playlist]
      responses:
        '200':
          description: Qobuz search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QobuzSearchResults'
...
  /api/downloads/qobuz:
    post:
      tags: [downloads]
      summary: Download from Qobuz
      operationId: downloadQobuz
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QobuzDownloadRequest'
      responses:
        '202':
          description: Download started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Download'
  /api/downloads/url:
    post:
      tags: [downloads]
      summary: Download from URL (YouTube, Bandcamp, etc.)
      operationId: downloadUrl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlDownloadRequest'
      responses:
        '202':
          description: Download started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Download'
...
  /api/downloads/queue:
    get:
      tags: [downloads]
      summary: Get download queue
      operationId: getDownloadQueue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active downloads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Download'
  /api/downloads/{id}:
    delete:
      tags: [downloads]
      summary: Cancel download
      operationId: cancelDownload
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Download cancelled

  # ==========================================================================
  # ADMIN
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ExportRequest:
      type: object
      required: [destination]
      properties:
        destination:
          type: string
        format:
          type: string
          enum: [flac, mp3, both]
          default: flac
        include_artwork:
          type: boolean
          default: true
        include_playlist:
          type: boolean
          default: false
    ExportStatus:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, running, complete, failed]
        progress:
          type: integer
        total_albums:
          type: integer
        exported_albums:
          type: integer
    QobuzSearchResults:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              artist:
                type: string
              year:
                type: integer
              url:
                type: string
              thumbnail:
                type: string
...
    QobuzDownloadRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Qobuz album/track URL
        search_type:
          type: string
          enum: [artist, album, track, playlist]
          description: Used for auto-heart rule (track only)
    UrlDownloadRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
        confirm_lossy:
          type: boolean
          default: false
          description: Required for YouTube/lossy sources
        search_type:
          type: string
          enum: [artist, album, track, playlist]
          description: Used for auto-heart rule (track only)
...
    Download:
      type: object
      properties:
        id:
          type: integer
        source:
          type: string
        search_type:
          type: string
          enum: [artist, album, track, playlist]
        status:
          type: string
          enum: [pending, downloading, importing, complete, failed, cancelled]
        progress:
          type: integer
        speed:
          type: string
        eta:
          type: string
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
...